requirements.txt
dearpygui
pillow
imagehash
